updated paths to solr codebase
Hey!

not sure if I misunderstood something obvious, but I do think the paths as updated by me were wrongly typed. The "with_items" loop had hardcoded stuff in it, so definitely spurious. I'm just testing it on my install so might report more, but so far these are the most obvious blockers.
I had mine ansible for solr but your looks more thorough and done again 4.x version, so thanks for that.