Use default Java versions when testing Solr 3/4 (except debian > 8)
Testing Solr on the newer versions of the OSes breaks the Java
installation on OSes that do not supported the older version of the
package. Since the java role has sensible defaults for various OS
versions this simply doesn't set a version so the default is used -
except debian 9. Debian 9 doesn't correct get java 8 but 7 is attempted
to be installed. So the test still overrides the java package for >
debian 8.

The apt cache needs to be updated as well due to a behavior difference
in the newer version of apt and ansible. It appears that ansible
currently incorrectly updates the apt cache in earlier OS versions so
the java installs wwere working correctly. In the newer versions,
however, trying to install java results in a package not found error due
to the empty cache.